Update for 2024 undirected instances

+++ src/graph.h
@@ -35,7 +35,7 @@ namespace fpc {
 
 typedef uint16_t Vertex;
 typedef std::pair<Vertex, Vertex> Edge;
-typedef uint8_t Edge_length;
+typedef uint16_t Edge_length;
 typedef mpz_class Edge_weight;
 typedef std::pair<Edge_length, Edge_weight> Weight;
 typedef std::priority_queue<std::pair<Edge_length, Vertex>,
